requirements.txt
alembic==1.13.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asyncpg==0.29.0
autoflake==2.3.1
bcrypt==4.2.0
billiard==4.2.1
black==25.1.0
celery==5.4.0
certifi==2024.6.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.111.0
fastapi-cache2==0.2.2
fastapi-cli==0.0.4
fastapi-versioning==0.10.0
flake8==7.1.1
flower==2.0.1
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanize==4.11.0
idna==3.7
iniconfig==2.0.0
isort==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
nodeenv==1.9.1
orjson==3.10.3
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
pendulum==3.0.0
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus-fastapi-instrumentator==7.0.2
prometheus_client==0.21.1
prompt_toolkit==3.0.48
pyasn1==0.6.0
pycodestyle==2.12.1
pydantic==2.7.3
pydantic-extra-types==2.8.0
pydantic-settings==2.3.1
pydantic_core==2.18.4
pyflakes==3.2.0
Pygments==2.18.0
pyright==1.1.393
pytest==8.3.4
pytest-asyncio==0.25.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==3.2.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.1
redis==4.6.0
rich==13.7.1
rsa==4.9
sentry-sdk==2.20.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqladmin==0.20.1
SQLAlchemy==2.0.32
starlette==0.37.2
time-machine==2.16.0
tornado==6.4.2
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.30.1
vine==5.1.0
watchdog==6.0.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
WTForms==3.1.2
